Disable handlers for UnobservedTaskException by default
Since .NET 4.5, by default, UnobservedTaskExceptions do no longer cause
the app to crash. The SDK has not been adapted for this and still logs
these errors and causes the program to exit, though this might not be
needed or intended.

This commit changes the default behavior to NOT add any handlers for
UnobservedTaskException.

Users who wish to continue using the handler, should add calls to
`RegisterUnobservedTaskExceptionHandler()` or
`RegisterDefaultUnobservedTaskExceptionHandler()`  after calling
`Configure()`.

Fixes #29.
Fixes #27 for other platforms than WPF.